Moved the depthai modules and tested the gesture collection script
