+++ examples/semaphore_alphabet/demo_semaphore.py
